refactor: datetime processing logic for i90 and i3 outsourced to date utils file for reuse
